Rename files back to their original names
