+++ test/unit/decorators/test_resources_decorator.py
@@ -1,6 +1,7 @@
 import pytest
 
 from layer.decorators import resources
+from layer.decorators.definitions import ResourcePath
 
 
 def test_resources_decorator_combines_all_paths():
@@ -8,7 +9,11 @@ def test_resources_decorator_combines_all_paths():
     def func():
         pass
 
-    assert func.layer.get_resource_paths() == ["/", "./abc", "../xyz"]
+    assert func.layer.get_resource_paths() == [
+        ResourcePath(path="/"),
+        ResourcePath(path="./abc"),
+        ResourcePath(path="../xyz"),
+    ]
 
 
 def test_resources_decorator_requires_at_least_one_path():
